log repeats files that are at the folder passed to the arguments. It repeats this as many times as there are subfolders at the folder specified at the arguments

+++ log.h
@@ -1,9 +1,10 @@
 #ifndef LOG_H
 #define LOG_H
 
+
 void start(const char* filename);
 
-long double getTime(void);
+double getTime(void);
 
 int logPrint(const char* output);
